Convert prerender script to Node.js
