+++ package.json
@@ -83,7 +83,7 @@
   },
   "scripts": {
     "build": "webpack --config webpack.config.production.js",
-    "prerender": "SCRIVITO_PRERENDER=true npm run build && node prerender/storePrerenderedContent.js",
+    "prerender": "SCRIVITO_PRERENDER=true npm run build && node buildPrerendering/prerender_content.js",
     "start": "webpack serve",
     "eslint": "eslint --max-warnings 0 *.js src/ prerender/",
     "es-check": "npm run build && es-check es5 --not=build/assets/scrivito_extensions. './build/**/*.js'"
