+++ src/prerender_content.js
@@ -1,11 +1,15 @@
 /* eslint no-console: "off" */
-import "scrivito";
-import filesize from "filesize";
+/* eslint-disable-next-line import/no-extraneous-dependencies */
+import path from "path";
+import fse from "fs-extra";
 import "./Objs";
 import "./Widgets";
 import { configure } from "./config";
 import prerenderObjs from "./prerenderContent/prerenderObjs";
 import prerenderSitemap from "./prerenderContent/prerenderSitemap";
+import { extendRedirects } from "./prerenderContent/extendRedirects";
+import { reportError } from "./prerenderContent/reportError";
+import { storeResult } from "./prerenderContent/storeResult";
 
 configure({ priority: "background" });
 
@@ -27,33 +31,44 @@ const SITEMAP_OBJ_CLASSES_WHITELIST = [
   "Page",
 ];
 
-async function prerenderContent() {
-  await prerenderSitemap(SITEMAP_OBJ_CLASSES_WHITELIST, window.storeResult);
+const SOURCE_DIR = "build";
+const TARGET_DIR = "buildPrerendered";
 
-  const assetManifest = await (await fetch("asset-manifest.json")).json();
+async function prerenderContent() {
+  console.time("[prerenderContent]");
+
+  const assetManifest = await fse.readJson(
+    path.join(SOURCE_DIR, "asset-manifest.json")
+  );
+
+  console.log(`Removing ${TARGET_DIR}/`);
+  await fse.remove(TARGET_DIR);
+
+  console.log(`Copying ${SOURCE_DIR}/ to ${TARGET_DIR}/`);
+  await fse.copy(SOURCE_DIR, TARGET_DIR);
+
+  await fse.remove(path.join(TARGET_DIR, "asset-manifest.json"));
+
+  const storedFiles = [];
+  const storeFile = (file) => storeResult(TARGET_DIR, storedFiles, file);
+
+  await prerenderSitemap(SITEMAP_OBJ_CLASSES_WHITELIST, storeFile);
+
   await prerenderObjs(
     PRERENDER_OBJ_CLASSES_BLACKLIST,
-    window.storeResult,
-    window.reportError,
+    storeFile,
+    reportError,
     assetManifest
   );
+
+  await extendRedirects(TARGET_DIR, storedFiles, SOURCE_DIR);
+
+  console.log(`📦 Added ${storedFiles.length} files to ${TARGET_DIR}!`);
+
+  console.timeEnd("[prerenderContent]");
 }
 
-// The following method will be overwritten by puppeteer in storePrerenderedContent.
-// It is only here, to simplify debugging in the browser
-window.storeResult = async ({ filename, content }) => {
-  console.log(
-    `[storeResult] received "${filename}" (file size: ${filesize(
-      content.length
-    )})`
-  );
-};
-
-// The following method will be overwritten by puppeteer in storePrerenderedContent.
-// It is only here, to simplify debugging in the browser
-window.reportError = (message, ...args) => {
-  console.log(`[reportError] ${message}`, ...args);
-};
-
-// Usage: window.prerenderContent().then(results => ...);
-window.prerenderContent = prerenderContent;
+prerenderContent().catch((e) => {
+  reportError("An error occurred!", e);
+  process.exitCode = 1;
+});
